services/order-service/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.31.0
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
uvicorn==0.27.0